refactor(Structure): provide helper methods for lists and dictionaries
There are a number of places where the use of a List or Dictionary has
unnecessary repetition in checking if a value or key is present before
either adding or removing from the list or dictionary.

There are also times where a value is attempted to be retrieved from
a dictionary but it doesn't exist which would then need to be added
so multiple dictionary lookups are done, which is inefficient. The
proposed changes ensure a single look up is done via `TryGetValue`
which will reduce dictionary usage overhead.